requirements.txt
absl-py==2.0.0
astunparse==1.6.3
blinker==1.7.0
cachecontrol==0.14.0
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.1.0
cryptography==41.0.3
cycler==0.11.0
faiss-cpu==1.7.4
filelock==3.12.2
firebase-admin==6.0.0
flask==2.3.2
flatbuffers==23.5.26
fonttools==4.40.0
fsspec==2023.6.0
gast==0.4.0
git-filter-repo==2.38.0
google-api-core==2.11.0
google-api-python-client==2.94.0
google-auth-httplib2==0.1.0
google-auth==2.22.0
google-cloud-core==2.3.2
google-cloud-firestore==2.11.1
google-cloud-storage==2.10.0
google-crc32c==1.5.0
google-pasta==0.2.0
google-resumable-media==2.6.0
googleapis-common-protos==1.59.0
grpcio-status==1.56.2
grpcio==1.56.2
gunicorn==20.1.0
h5py==3.9.0
httplib2==0.22.0
huggingface-hub==0.16.4
idna==3.4
imbalanced-learn==0.10.1
imblearn==0.0
itsdangerous==2.1.2
jinja2==3.1.2
joblib==1.2.0
keras==2.12.0
kiwisolver==1.4.4
libclang==16.0.6
markdown-it-py==2.2.0
markdown==3.4.4
markupsafe==2.1.3
matplotlib==3.7.1
mdurl==0.1.2
ml-dtypes==0.2.0
mpmath==1.3.0
msgpack==1.0.5
namex==0.0.9
networkx==3.1
nltk==3.8.1
numpy==1.23.5
opt-einsum==3.3.0
optree==0.10.0
packaging==23.1
pandas==1.5.3
pillow==9.5.0
proto-plus==1.22.3
protobuf==4.23.4
pyasn1-modules==0.3.0
pyasn1==0.5.0
pycparser==2.21
pygments==2.15.1
pyjwt==2.7.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
pyyaml==6.0
regex==2023.6.3
requests==2.31.0
rich==13.4.2
rsa==4.9
safetensors==0.3.3
sastrawi==1.0.1
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
sklearn-compat==0.1.3
sympy==1.12
tensorboard-data-server==0.7.2
tensorboard==2.11.2
tensorflow-cpu==2.12.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
tf-keras==2.11.0
threadpoolctl==3.1.0
tokenizers==0.13.3
torch==2.0.1
tqdm==4.65.0
transformers==4.30.2
typing-extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.16
werkzeug==2.3.6
wordcloud==1.9.2
wrapt==1.15.0